Google-The_Gemma_3n_Impact_Challenge/PocketDoc/requirements.txt
torch
transformers
accelerate
streamlit
Pillow
